+++ e2e/tests/functional/planning/timelistControlledClock.e2e.spec.js
@@ -28,7 +28,8 @@ clockOptions plugin fixture.
 import fs from 'fs';
 
 import { createDomainObjectWithDefaults, createPlanFromJSON } from '../../../appActions.js';
-import { getEarliestStartTime } from '../../../helper/planningUtils.js';
+import { getEarliestStartTime, getFirstActivity } from '../../../helper/planningUtils';
+
 import { expect, test } from '../../../pluginFixtures.js';
 
 const examplePlanSmall3 = JSON.parse(
@@ -40,6 +41,9 @@ const examplePlanSmall3 = JSON.parse(
 const TIME_TO_FROM_COLUMN = 2;
 const HEADER_ROW = 0;
 const NUM_COLUMNS = 5;
+
+const FULL_CIRCLE_PATH =
+  'M3.061616997868383e-15,-50A50,50,0,1,1,-3.061616997868383e-15,50A50,50,0,1,1,3.061616997868383e-15,-50Z';
 
 /**
  * The regular expression used to parse the countdown string.
@@ -158,6 +162,119 @@ test.describe('Time List with controlled clock @clock', () => {
     }
   });
 });
+
+test.describe('Activity progress when activity is in the future @clock', () => {
+  const firstActivity = getFirstActivity(examplePlanSmall1);
+
+  test.use({
+    clockOptions: {
+      now: firstActivity.start - 1,
+      shouldAdvanceTime: true
+    }
+  });
+
+  test.beforeEach(async ({ page }) => {
+    await createTimelistWithPlanAndSetActivityInProgress(page);
+  });
+
+  test('progress pie is empty', async ({ page }) => {
+    const anActivity = page.getByRole('row').nth(0);
+    // Progress pie shows no progress when now is less than the start time
+    await expect(anActivity.getByLabel('Activity in progress').locator('path')).not.toHaveAttribute(
+      'd'
+    );
+  });
+});
+
+test.describe('Activity progress when now is between start and end of the activity @clock', () => {
+  const firstActivity = getFirstActivity(examplePlanSmall1);
+  test.beforeEach(async ({ page }) => {
+    await createTimelistWithPlanAndSetActivityInProgress(page);
+  });
+
+  test.use({
+    clockOptions: {
+      now: firstActivity.start + 50000,
+      shouldAdvanceTime: true
+    }
+  });
+
+  test('progress pie is partially filled', async ({ page }) => {
+    const anActivity = page.getByRole('row').nth(0);
+    const pathElement = anActivity.getByLabel('Activity in progress').locator('path');
+    // Progress pie shows progress when now is greater than the start time
+    await expect(pathElement).toHaveAttribute('d');
+  });
+});
+
+test.describe('Activity progress when now is after end of the activity @clock', () => {
+  const firstActivity = getFirstActivity(examplePlanSmall1);
+
+  test.use({
+    clockOptions: {
+      now: firstActivity.end + 10000,
+      shouldAdvanceTime: true
+    }
+  });
+
+  test.beforeEach(async ({ page }) => {
+    await createTimelistWithPlanAndSetActivityInProgress(page);
+  });
+
+  test('progress pie is full', async ({ page }) => {
+    const anActivity = page.getByRole('row').nth(0);
+    // Progress pie is completely full and doesn't update if now is greater than the end time
+    await expect(anActivity.getByLabel('Activity in progress').locator('path')).toHaveAttribute(
+      'd',
+      FULL_CIRCLE_PATH
+    );
+  });
+});
+
+async function createTimelistWithPlanAndSetActivityInProgress(page) {
+  await page.goto('./', { waitUntil: 'domcontentloaded' });
+
+  const timelist = await createDomainObjectWithDefaults(page, { type: 'Time List' });
+
+  await createPlanFromJSON(page, {
+    name: 'Test Plan',
+    json: examplePlanSmall1,
+    parent: timelist.uuid
+  });
+
+  // Ensure that all activities are shown in the expanded view
+  const groups = Object.keys(examplePlanSmall1);
+  const firstGroupKey = groups[0];
+  const firstGroupItems = examplePlanSmall1[firstGroupKey];
+  const firstActivityForPlan = firstGroupItems[0];
+  const lastActivity = firstGroupItems[firstGroupItems.length - 1];
+  const startBound = firstActivityForPlan.start;
+  const endBound = lastActivity.end;
+
+  // Switch to fixed time mode with all plan events within the bounds
+  await page.goto(
+    `${timelist.url}?tc.mode=fixed&tc.startBound=${startBound}&tc.endBound=${endBound}&tc.timeSystem=utc&view=timelist.view`
+  );
+
+  // Change the object to edit mode
+  await page.getByRole('button', { name: 'Edit Object' }).click();
+
+  // Find the display properties section in the inspector
+  await page.getByRole('tab', { name: 'View Properties' }).click();
+  // Switch to expanded view and save the setting
+  await page.getByLabel('Display Style').selectOption({ label: 'Expanded' });
+
+  // Click on the "Save" button
+  await page.getByRole('button', { name: 'Save' }).click();
+  await page.getByRole('listitem', { name: 'Save and Finish Editing' }).click();
+
+  const anActivity = page.getByRole('row').nth(0);
+
+  // Set the activity to in progress
+  await anActivity.click();
+  await page.getByRole('tab', { name: 'Activity' }).click();
+  await page.getByLabel('Activity Status', { exact: true }).selectOption({ label: 'In progress' });
+}
 
 /**
  * Get the cell at the given row and column indices.
